Move status bar code out of EditorManager into a new module. Keeps EditorManager tightly focused on maintaining the set of Editors and gives the status bar code room to grow for future UI additions.
This removes EditorManager._init() and the manual calls to it from unit
tests, which seem unnecessary. EditorStatusBar._init() is still called on
htmlReady - just not manually called by any unit tests.

Also, change the status bar code to use active editor more consistently
(used to use focused editor sometimes).
