+++ src/editor/EditorManager.js
@@ -100,14 +100,6 @@ define(function (require, exports, module) {
      */
     var _inlineEditProviders = [];
     
-    /* StatusBar indicators */
-    var $languageInfo,
-        $cursorInfo,
-        $fileInfo,
-        $indentType,
-        $indentWidthLabel,
-        $indentWidthInput;
-    
 	/**
      * @private
      * @param {?Editor} current
@@ -696,137 +688,6 @@ define(function (require, exports, module) {
         return result.promise();
     }
     
-    function _updateLanguageInfo(editor) {
-        $languageInfo.text(editor.document.getLanguage().getName());
-    }
-    
-    function _updateFileInfo(editor) {
-        $fileInfo.text(StringUtils.format(Strings.STATUSBAR_LINE_COUNT, editor.lineCount()));
-    }
-    
-    function _updateIndentType() {
-        var indentWithTabs = Editor.getUseTabChar();
-        $indentType.text(indentWithTabs ? Strings.STATUSBAR_TAB_SIZE : Strings.STATUSBAR_SPACES);
-        $indentType.attr("title", indentWithTabs ? Strings.STATUSBAR_INDENT_TOOLTIP_SPACES : Strings.STATUSBAR_INDENT_TOOLTIP_TABS);
-        $indentWidthLabel.attr("title", indentWithTabs ? Strings.STATUSBAR_INDENT_SIZE_TOOLTIP_TABS : Strings.STATUSBAR_INDENT_SIZE_TOOLTIP_SPACES);
-    }
-
-    function _getIndentSize() {
-        return Editor.getUseTabChar() ? Editor.getTabSize() : Editor.getIndentUnit();
-    }
-    
-    function _updateIndentSize() {
-        var size = _getIndentSize();
-        $indentWidthLabel.text(size);
-        $indentWidthInput.val(size);
-    }
-    
-    function _toggleIndentType() {
-        Editor.setUseTabChar(!Editor.getUseTabChar());
-        _updateIndentType();
-        _updateIndentSize();
-    }
-    
-    function _updateCursorInfo(event, editor) {
-        editor = editor || getFocusedEditor();
-
-        // compute columns, account for tab size
-        var cursor = editor.getCursorPos(true);
-        
-        $cursorInfo.text(StringUtils.format(Strings.STATUSBAR_CURSOR_POSITION, cursor.line + 1, cursor.ch + 1));
-    }
-    
-    function _changeIndentWidth(value) {
-        $indentWidthLabel.removeClass("hidden");
-        $indentWidthInput.addClass("hidden");
-        
-        // remove all event handlers from the input field
-        $indentWidthInput.off("blur keyup");
-        
-        // restore focus to the editor
-        focusEditor();
-
-        if (!value || isNaN(value)) {
-            return;
-        }
-        
-        if (Editor.getUseTabChar()) {
-            Editor.setTabSize(Math.max(Math.min(value, 10), 1));
-        } else {
-            Editor.setIndentUnit(Math.max(Math.min(value, 10), 1));
-        }
-
-        // update indicator
-        _updateIndentSize();
-
-        // column position may change when tab size changes
-        _updateCursorInfo();
-    }
-    
-    function _onActiveEditorChange(event, current, previous) {
-        if (previous) {
-            $(previous).off("cursorActivity.statusbar");
-            $(previous).off("change.statusbar");
-        }
-        
-        if (!current) {
-            StatusBar.hide();  // calls resizeEditor() if needed
-        } else {
-            StatusBar.show();  // calls resizeEditor() if needed
-            
-            $(current).on("cursorActivity.statusbar", _updateCursorInfo);
-            $(current).on("change.statusbar", function () {
-                // async update to keep typing speed smooth
-                window.setTimeout(function () { _updateFileInfo(current); }, 0);
-            });
-            
-            _updateCursorInfo(null, current);
-            _updateLanguageInfo(current);
-            _updateFileInfo(current);
-            _updateIndentType();
-            _updateIndentSize();
-        }
-    }
-    
-    function _init() {
-        StatusBar.init($(".main-view .content"));
-
-        $languageInfo       = $("#status-language");
-        $cursorInfo         = $("#status-cursor");
-        $fileInfo           = $("#status-file");
-        $indentType         = $("#indent-type");
-        $indentWidthLabel   = $("#indent-width-label");
-        $indentWidthInput   = $("#indent-width-input");
-        
-        // indentation event handlers
-        $indentType.on("click", _toggleIndentType);
-        $indentWidthLabel
-            .on("click", function () {
-                // update the input value before displaying
-                $indentWidthInput.val(_getIndentSize());
-
-                $indentWidthLabel.addClass("hidden");
-                $indentWidthInput.removeClass("hidden");
-                $indentWidthInput.focus();
-        
-                $indentWidthInput
-                    .on("blur", function () {
-                        _changeIndentWidth($indentWidthInput.val());
-                    })
-                    .on("keyup", function (event) {
-                        if (event.keyCode === KeyEvent.DOM_VK_RETURN) {
-                            $indentWidthInput.blur();
-                        } else if (event.keyCode === KeyEvent.DOM_VK_ESCAPE) {
-                            _changeIndentWidth(false);
-                        }
-                    });
-            });
-
-        $indentWidthInput.focus(function () { $indentWidthInput.select(); });
-
-        _onActiveEditorChange(null, getFocusedEditor(), null);
-    }
-
     // Initialize: command handlers
     CommandManager.register(Strings.CMD_TOGGLE_QUICK_EDIT, Commands.TOGGLE_QUICK_EDIT, _toggleQuickEdit);
     
@@ -839,13 +700,7 @@ define(function (require, exports, module) {
     // refresh on resize.
     window.addEventListener("resize", _updateEditorDuringResize, true);
     
-    // Initialize: status bar focused listener
-    $(exports).on("activeEditorChange", _onActiveEditorChange);
-    
-    AppInit.htmlReady(_init);
-    
     // For unit tests and internal use only
-    exports._init = _init;
     exports._openInlineWidget = _openInlineWidget;
     exports._createFullEditorForDocument = _createFullEditorForDocument;
     exports._destroyEditorIfUnneeded = _destroyEditorIfUnneeded;
